Fix procedures namespace
